+++ include/rudiments/private/clientsocket.h
@@ -6,7 +6,7 @@
 		void	*newSSLBIO() const;
 #endif
 		int32_t	connect(const struct ::sockaddr *addr,
-				socklen_t addrlen, long sec, long usec);
+				socklen_t addrlen, int32_t sec, int32_t usec);
 		ssize_t	lowLevelRead(void *buf, ssize_t count) const;
 		ssize_t	lowLevelWrite(const void *buf, ssize_t count) const;
 		int32_t	lowLevelClose();
