+++ include/rudiments/private/listener.h
@@ -2,7 +2,7 @@
 // See the COPYING file for more information.
 
 	protected:
-		virtual	int32_t	safeSelect(long sec, long usec,
+		virtual	int32_t	safeSelect(int32_t sec, int32_t usec,
 						bool read, bool write);
 
 		// FIXME: it should be possible to copy a listener
